% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.vars.R
\name{arrange.vars}
\alias{arrange.vars}
\title{Rearrange Columns in a Data Frame}
\usage{
arrange.vars(data, vars)
}
\arguments{
\item{data}{A data frame in which columns need to be rearranged.}

\item{vars}{A named vector specifying the new positions of the columns. The names of the vector should be the column names, and the values should be the desired positions.}
}
\value{
A data frame with the columns rearranged according to the specified order.
}
\description{
This function rearranges the columns of a data frame based on a specified order.
}
\details{
The function takes a data frame and a named vector as input. The named vector indicates the desired column names and their new positions. It performs sanity checks to ensure that the specified columns exist in the data frame, that the positions are within valid range, and that there are no duplicates. The function returns the data frame with columns rearranged as specified.
}
