% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codon_usage_calculator.R
\name{Codon_usage_calculator}
\alias{Codon_usage_calculator}
\title{codon usage table}
\usage{
Codon_usage_calculator(target = NULL, save_output = NULL)
}
\arguments{
\item{df}{A data frame containing the data from the genbank_organizer function in DNMB package.}
}
\value{
A data frame containing the codon usage information.
}
\description{
codon usage table
}
