% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage_tRNA_heatmap_generator.R
\name{codon_usage_tRNA_heatmap_generator}
\alias{codon_usage_tRNA_heatmap_generator}
\title{Generate a heatmap for codon usage and tRNA copy number.}
\usage{
codon_usage_tRNA_heatmap_generator(
  codon_usage = NULL,
  tRNA_distribution = NULL,
  output_pdf = "codon_usage_tRNA_heatmap.pdf",
  width = 10,
  height = 10,
  save_pdf = FALSE
)
}
\arguments{
\item{codon_usage}{A data frame containing codon usage data.}

\item{tRNA_distribution}{A data frame containing tRNA copy number data.}

\item{output_pdf}{A string specifying the name of the output PDF file.}

\item{width}{A numeric value specifying the width of the output PDF in inches.}

\item{height}{A numeric value specifying the height of the output PDF in inches.}

\item{save_pdf}{A logical value indicating whether to save the plot as a PDF (default is TRUE).}
}
\value{
A heatmap figure combining codon usage and tRNA copy number information.
}
\description{
This function creates a heatmap that visualizes codon usage and the corresponding tRNA copy number distribution.
}
