% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPR_array_extractor.R
\name{CRISPR_array_extractor}
\alias{CRISPR_array_extractor}
\title{CRISPR array extractor}
\usage{
CRISPR_array_extractor(gb_dir = NULL, save_output = FALSE)
}
\arguments{
\item{gb_dir}{A string specifying the directory where the InterProScan output files are located. If NULL, the current working directory is used.}

\item{df}{A data frame containing the data from the InterProScan (tsv format).}
}
\value{
A data frame containing the baseline table.
}
\description{
CRISPR array extractor
}
