% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cazy_carbon.R
\name{.dnmb_cct_3zone_extract_gh}
\alias{.dnmb_cct_3zone_extract_gh}
\title{Extract GH enzymes with locus_tag, gene_name, family from genbank_table}
\usage{
.dnmb_cct_3zone_extract_gh(genbank_table)
}
\description{
Extract GH enzymes with locus_tag, gene_name, family from genbank_table
}
\keyword{internal}
