% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cazy_carbon.R
\name{.dnmb_cct_3zone_extract_transporters}
\alias{.dnmb_cct_3zone_extract_transporters}
\title{Extract transporter genes from genbank_table (GapMindCarbon transport steps)}
\usage{
.dnmb_cct_3zone_extract_transporters(genbank_table, output_dir = NULL)
}
\description{
Extract transporter genes from genbank_table (GapMindCarbon transport steps)
}
\keyword{internal}
