% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cazy_carbon.R
\name{.dnmb_scissors_grob_v2}
\alias{.dnmb_scissors_grob_v2}
\title{Draw scissors symbol (two crossed blades) for GH cleavage}
\usage{
.dnmb_scissors_grob_v2(x, y, size = 0.12, angle = 0)
}
\arguments{
\item{x, y}{Position of the scissors}

\item{size}{Scale factor}

\item{angle}{Rotation angle in degrees (0 = horizontal cut)}
}
\value{
List of ggplot2 layers (two curved blade segments + pivot)
}
\description{
Draw scissors symbol (two crossed blades) for GH cleavage
}
\keyword{internal}
