% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure_generator.R
\name{Figure_generator}
\alias{Figure_generator}
\title{Generate a complex plot with heatmap, seqlogo, and histogram}
\usage{
Figure_generator(
  output_pdf = "arranged_plot_with_labels.pdf",
  pdf_width = 12,
  pdf_height = 12
)
}
\arguments{
\item{output_pdf}{The output path for the PDF file.}

\item{pdf_width}{The width of the output PDF file (default is 12).}

\item{pdf_height}{The height of the output PDF file (default is 12).}
}
\value{
A PDF file with the arranged plots and labels.
}
\description{
Generate a complex plot with heatmap, seqlogo, and histogram
}
