% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genome_summary.R
\name{gb_info}
\alias{gb_info}
\title{genbank feature table}
\usage{
gb_info(gb_dir = NULL)
}
\arguments{
\item{gb_dir}{A genbank file containing nucleotide such as ".gb", "gbk", ".gbff". the directory where the genbank files are located. If NULL, the current working directory is used.}
}
\value{
A data frame from genbank file.
}
\description{
This is a simple example function.
}
