% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genbank_fna_extractor.R
\name{genbank_fna_extractor}
\alias{genbank_fna_extractor}
\title{Genbank fna extractor}
\usage{
genbank_fna_extractor(gb_dir = NULL)
}
\arguments{
\item{gb_dir}{A string specifying the directory where the genbank output files are located. If NULL, the current working directory is used.}

\item{df}{A data frame containing the data from the genbank.}
}
\value{
A data frame containing the baseline table.
}
\description{
Genbank fna extractor
}
