% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genbank_organizer.R
\name{Genbank_organizer}
\alias{Genbank_organizer}
\title{GenBank organizer}
\usage{
Genbank_organizer(gb_dir = NULL, save_output = FALSE)
}
\arguments{
\item{gb_dir}{A string specifying the directory where the genbank output files are located. If NULL, the current working directory is used.}

\item{df}{A data frame containing the data from the genbank.}
}
\value{
A data frame containing the baseline table.
}
\description{
GenBank organizer
}
