% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterProScan_organizer.R
\name{InterProScan_annotations}
\alias{InterProScan_annotations}
\title{InterProScan table}
\usage{
InterProScan_annotations(InterProScan_dir = NULL)
}
\arguments{
\item{InterProScan_dir}{A string specifying the directory where the InterProScan output files are located. If NULL, the current working directory is used.}

\item{df}{A data frame containing the data from the InterProScan (tsv format).}
}
\value{
A data frame containing the baseline table.
}
\description{
This is a simple example function.
}
