% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBS_extractor.R
\name{RBS_extractor}
\alias{RBS_extractor}
\title{Extract the RBS (Ribosome Binding Site) region.}
\usage{
RBS_extractor(
  target = NULL,
  plot_RBS = FALSE,
  save_plot = FALSE,
  plot_path = NULL
)
}
\arguments{
\item{target}{A data frame containing the target data for analysis.}

\item{plot_RBS}{A logical value indicating whether to generate a plot of the RBS region.}

\item{plot_path}{A string specifying the file path to save the RBS plot (if \code{plot_RBS} is TRUE).}

\item{dependent}{A string specifying the dependent variable for the analysis.}
}
\value{
A data frame containing the extracted RBS regions.
}
\description{
This function extracts the Ribosome Binding Site (RBS) region from the target data frame.
}
