% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Promotech.R
\name{save_and_run_promotech}
\alias{save_and_run_promotech}
\title{Save Contig Sequences and Run External Tool}
\usage{
save_and_run_promotech(
  conda_env_path,
  promotech_script_path,
  output_dir = "results"
)
}
\arguments{
\item{conda_env_path}{A character string specifying the path to the conda environment
(e.g., \verb{/Users/JaeYoon/miniforge3/envs/promotech_mac_env}).}

\item{promotech_script_path}{A character string specifying the path to the \code{promotech.py} script.}

\item{output_dir}{A character string specifying the directory to store the results. Default is \code{results}.}
}
\value{
This function does not return a value but saves \code{.fna} files and runs \code{promotech.py}.
Outputs from the external tool are stored in the specified results folder.
}
\description{
This script identifies objects in the R environment that match the pattern \verb{contig_[number]_seq},
saves them as FASTA files, and runs an external Python tool on each file using a conda environment.
}
\details{
The function performs the following steps:
\enumerate{
\item Identifies all objects in the R environment matching the pattern \verb{contig_[number]_seq}.
\item Saves each matching object as a \code{.fna} file in FASTA format.
\item Executes the external Python tool \code{promotech.py} on each \code{.fna} file using a conda environment.
\item Outputs the results into a designated \code{results} folder.
}
}
\examples{
# Example usage:
save_and_run_promotech(
  conda_env_path = "/Users/JaeYoon/miniforge3/envs/promotech_mac_env",
  promotech_script_path = "~/Desktop/Bioinformatics/promotech-master/promotech.py"
)

}
