% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count_tRNA_anticodon.R
\name{tRNA_anticodon_counter}
\alias{tRNA_anticodon_counter}
\title{Count tRNA anticodon}
\usage{
tRNA_anticodon_counter(target = NULL, save_output = NULL)
}
\arguments{
\item{df}{A data frame containing the data from the genbank.}

\item{gb_dir}{A string specifying the directory where the genbank output files are located. If NULL, the current working directory is used.}
}
\value{
A data frame containing the tRNA coding region and distribution.
}
\description{
Count tRNA anticodon
}
