% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPR_array_extractor.R
\name{CRISPR_array_extractor}
\alias{CRISPR_array_extractor}
\title{CRISPR array extractor}
\usage{
CRISPR_array_extractor(gb_dir = NULL, save_output = FALSE)
}
\arguments{
\item{gb_dir}{A string specifying the directory where the GenBank files are located. If NULL, the current working directory is used.}

\item{save_output}{Logical, whether to save the output to Excel and FASTA files.}
}
\value{
A data frame containing the extracted CRISPR array information.
}
\description{
CRISPR array extractor
}
