% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_combiner.R
\name{DNMB_table}
\alias{DNMB_table}
\title{Result DNMB table}
\usage{
DNMB_table(
  genbank_table = TRUE,
  EggNOG_table = FALSE,
  InterProScan_table = FALSE,
  InterProScan_site = FALSE,
  codon_usage = TRUE,
  tRNA_anticodon = TRUE,
  tRNA_distribution = TRUE,
  RBS_table = TRUE,
  CRISPR_by_spacer = TRUE,
  save_dir = NULL
)
}
\arguments{
\item{gb_table}{A data frame containing the data.}

\item{InterPro_search}{Logical, whether to include p-values for means.}

\item{InterPro_site}{Logical, whether to include p-values for medians.}

\item{codon}{Logical, whether to use chi-square test for categorical variables.}
}
\value{
A data frame containing the baseline table.
}
\description{
Result DNMB table
}
