% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Promotech_genbank.R
\name{Promoter_to_genbank}
\alias{Promoter_to_genbank}
\title{Promotech to sanpgene #}
\usage{
Promoter_to_genbank(promotech = NULL, genbank = NULL)
}
\arguments{
\item{df}{A data frame containing the data from eggNOG-mapper (http://eggnog-mapper.embl.de).}

\item{EggNOG_dir}{A string specifying the directory where the eggNOG-mapper output files are located. If NULL, the current working directory is used.}

\item{save_output}{A logical value indicating whether to save the output to an Excel file in XLSX format. Defaults to FALSE.}
}
\value{
A list containing summary information of the processed data.
}
\description{
Promotech to sanpgene #
}
