% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNMB_combiner.R
\name{run_DNMB_combiner}
\alias{run_DNMB_combiner}
\title{Result DNMB table}
\usage{
run_DNMB_combiner(
  genbank_table = NULL,
  EggNOG_table = NULL,
  InterProScan_table = NULL
)
}
\arguments{
\item{gb_table}{A data frame containing the data.}

\item{InterPro_search}{Logical, whether to include p-values for means.}

\item{InterPro_site}{Logical, whether to include p-values for medians.}

\item{codon}{Logical, whether to use chi-square test for categorical variables.}
}
\value{
A data frame containing the baseline table.
}
\description{
Result DNMB table
}
